projects/capstone/starter/requirements.txt
alembic==1.4.2
appdirs==1.4.3
Authlib==0.14.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
distlib==0.3.0
ecdsa==0.15
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
pipenv==2018.11.26
psycopg2-binary
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jose==3.1.0
requests==2.23.0
rsa==4.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.1
